Fixed compiler warnings
* Missing function prototype in fts3_unicode2.c
* Implicit 64-to-32-bit conversions, when building for 64-bit CPU

+++ fts3_unicodesn.c
@@ -389,7 +389,7 @@ static int unicodeCreate(
 
   for(i=0; rc==SQLITE_OK && i<nArg; i++){
     const char *z = azArg[i];
-    int n = strlen(z);
+    int n = (int)strlen(z);
 
     if( n==19 && memcmp("remove_diacritics=1", z, 19)==0 ){
       pNew->bRemoveDiacritic = 1;
@@ -540,10 +540,10 @@ static int unicodeNext(
   );
 
   if ( pCsr->pStemmer!=NULL ) {
-     SN_set_current(pCsr->pStemmer, zOut - pCsr->zToken, (unsigned char *)pCsr->zToken);
+     SN_set_current(pCsr->pStemmer, (int)(zOut - pCsr->zToken), (unsigned char *)pCsr->zToken);
      if ( p->stemmer.stem(pCsr->pStemmer)<0 ) {
 	*paToken = pCsr->zToken;
-	*pnToken = zOut - pCsr->zToken;
+	*pnToken = (int)(zOut - pCsr->zToken);
      }else {
 	pCsr->pStemmer->p[pCsr->pStemmer->l] = '\0';
 	*paToken = (char *)pCsr->pStemmer->p;
@@ -551,13 +551,13 @@ static int unicodeNext(
      }
   }else {
      *paToken = pCsr->zToken;
-     *pnToken = zOut - pCsr->zToken;
+     *pnToken = (int)(zOut - pCsr->zToken);
   }
 
   /* Set the output variables and return. */
-  pCsr->iOff = (z - pCsr->aInput);
-  *piStart = (zStart - pCsr->aInput);
-  *piEnd = (zEnd - pCsr->aInput);
+  pCsr->iOff = (int)(z - pCsr->aInput);
+  *piStart = (int)(zStart - pCsr->aInput);
+  *piEnd = (int)(zEnd - pCsr->aInput);
   *piPos = pCsr->iToken++;
   return SQLITE_OK;
 }
